// src/app/models/time.series.ts
 export class  TimeSeriesModel {
    Units: number;
    Year: number;
    Month: number;
    Day: number;
    MeanTemp: number;
    Day_Of_the_Week: number;
    Week_of_the_Month: number;
    Week_of_the_Year: number;
    IsWeekDay: number;
    Winter_State: number;
    Holidays: number;
    Event_SchoolHoliday: number;
    is_month_start: number;
    is_month_end: number;
    quarter: number;
    is_quarter_start: number;
    is_quarter_end: number;
    is_year_end: number;
    is_year_start: number;
    is_leap_year: number;
    rolling_mean: number;
    expanding_mean: number;
    UnitsLag_0: number;
    UnitsLag_2: number;
    UnitsLag_5: number;
    UnitsLag_8: number;
    UnitsLag_11: number;
    UnitsLag_12: number;
    MW_Prediction: number;
    Date: string;
    Cat:string;
}